VSCSIMapping.bld_from_existing to clone storage
Discovered when attempting to map the same disk to more than one
partition.  If VSCSIMapping.bld_from_existing is given a storage element
wrapper that refers to an element residing in a chunk of XML, it pulls
that subtree out of that XML.  This is only a problem if that chunk of
XML is going to be used again - as it was in this case, where it
belonged to a mapping that we wanted to preserve.

Fixed to clone the storage element as well as the surrounding mapping.

Change-Id: I3c35db32c1444ed63949b88218b7902cc4690191