Refactor parse of latest metric data
The previous 'parse' method was very useful at first determining the
latest feed for phyp and VIOS metrics, and then parsing them down into
VM metrics.  This was used by the metric cache.

However, the activity of finding the 'latest' metrics is not exclusive
to VM metrics.  If users wish to derive host stats as a single entity,
the VM stats are just a piece.

That use case still requires that we find the latest stats, but instead
don't parse down to the VM scoped metrics...just return the raw PHYP
metrics back instead.

This change set refactors the existing code so that there is a new
method that returns the latest 'raw' stats.  The LparCache is then
refactored to use this common method.

Change-Id: I20b8a77cda826d1486677500b0a20c5e8188f27a