Ability to set Wrapper UUID
New Wrapper method _set_uuid allowing the UUID of a wrapper XML object
to be set.  For atom entries, roperly handles setting both the atom's
property and the Metadata/Atom/AtomID of the atom's content.  Properly
handles vivifying the (optional) AtomID child when it doesn't exist.
Refuses to operate on an element with no metadata.

Change-Id: I9e7303a405ade2854a3e31b1e33780f8b039c564