Separate metric cache concern
The pattern to gather the metrics once every 30 seconds is not unique to
the LPAR metrics.  This change set separates the cache into a separate
cache that other implementations can make use of.

The new mechanism also provides a means to not query the VIOSes (which
can be expensive).  This would not be needed for scenarios that only
need PHYP data (ex. gathering just CPU data).

This also fixes a bug that was identified in the previous metric
gathering.

Change-Id: I69edc1a7dacda70bab8982cf9080144bfe77812d