VSCSIMapping.bld_from_existing ignore TargetDevice
When "fusing" SCSI mappings using VSCSIMapping.bld_from_existing, we
were previously cloning the entire mapping, including the TargetDevice.
The TargetDevice is generated by the system and is specific to the
Storage element.  It is therefore incorrect to clone this piece.

This change set alters VSCSIMapping.bld_from_existing to clone the
mapping piecemeal.

Note: We could have opted to keep the existing cloning and explicitly
clear the TargetDevice field; this approach was chosen for best forward
compatibility (in case other VIOS-generated fields were later added to
the schema).

Change-Id: I841d72ce24d882d11bb5c3fe47154cd11d9f3739