Allow the uuid to be set via lpar builder
When lpars are created, PowerVM will automatically assign a uuid
value. Recent versions of PowerVM allow the uuid to be assigned
by the caller as long as it conforms to the PowerVM format.
Care should be taken when assigning a uuid value to ensure no
other lpar exists with the same uuid.

Change-Id: Id7c26a47ef8071bf45f6b1680129b3bce1c7f431