Update datetime formatter for LTM feed
New versions of the PowerVM REST API have fixed a minor bug in the date
time for the LTM metric feed.  It now returns the time zone.

This change set updates the wrappers to parse out the time zone, while
also supporting older versions of the format for previous REST API
levels.

The timezone is actually removed completely, as the datetime library has
significant issues parsing the timezone if there is no timezone
database.  Given the usage of how the timestamps are used in metrics,
this is perfectly acceptable as they are mainly used to sort for the
order of the metric responses.

Change-Id: Idbc0c62b881cdd3d4ffc9b701fe8d4d5870e3d0d