Update datetime formatter for LTM feed
New versions of the PowerVM REST API have fixed a minor bug in the date
time for the LTM metric feed.  It now returns the time zone.

This change set updates the wrappers to parse out the time zone, while
also supporting older versions of the format for previous REST API
levels.

Change-Id: Idbc0c62b881cdd3d4ffc9b701fe8d4d5870e3d0d