Tolerate any version of psutil
Versions of pypowervm prior to this change set required psutil at a
level prior to the conversion of Process.cmdline and .terminal from
@propertys to regular methods.  In situations where pip is used to pull
dependencies, this is not a problem, because pip seems to be able to
find older versions.  However, when pypowervm is installed from apt, the
psutil dependency is retrieved via apt (as python-psutil) from whatever
repository is configured on the platform.  On Ubuntu versions prior to
15, this is 1.2.1, which worked fine.  On 15, the default repositories
find 2.1.1-1 (as of this writing) and cannot find older versions.
Special machinations are therefore required to install and run pypowervm
on newer Ubuntu systems.

This change set removes the restriction to older versions of psutil by
creating a shim layer that behaves like the old version regardless of
what's installed.

Change-Id: I30de6bfe3a596826cf97efc25896e5d970b5b168