find_maps can include orphaned mappings
pypowervm.tasks.scsi_mapper.find_maps would previously ignore any
mappings considered "orphans" - that is, containing a server adapter but
no client adapter.  In cases where stale mappings (those associated with
a client LPAR that doesn't exist) need to be cleaned up, it is important
to be able to find these orphans so that, when the stale LPAR's ID is
reused, those mappings don't interfere with the new partition.

This change set adds an optional parameter to find_maps allowing the
consumer to specify that orphans should be included in the search.

Change-Id: I836f2d79a8e610ace8ea4e5d3037814a6ccbfc8d