Index VSCSI Mappings by LPAR, storage, target
New method pypowervm.tasks.scsi_mapper.index_mappings which indexes a
list of SCSI mappings so they can be looked up according to the client
LPAR's ID or UUID, or the UDID of the backing storage or target device.

This is to facilitate cleanup of storage mappings/artifacts associated
with a defunct LPAR.  For instance, we want to be able to iterate over
the mappings for a given LPAR, and for each mapping, determine whether
its storage and/or target device is associated with any mappings other
than this one; and if so, whether those mappings belong to the same LPAR
or a different one.

Change-Id: I6a15084d20b78282cfe94b4a6550ee7bef31dcdd