Allow for setting of io_slots on BasePartition wrapper
This allows for the removal or creation of ProfileIOSlots on a Partition
through the call to update on that Partition wrapper after setting the
ProfileIOSlots.

For example, removing a ProfileIOSlot from an existing Partition so that
slot can be added to a partition to be created is accomplished by this change.

Change-Id: I990e70e98f46000793712d4d5d392de727bd1b74