UseNextAvailableHighSlotID for CNA & TrunkAdapter
The PowerVM REST API recently implemented a new VirtualIOAdapter
property called UseNextAvailableHighSlotID.  This is used the same way
as the existing UseNextAvailableSlotID - to ask the PowerVM API to
assign a slot ID to a virtual adapter rather than using an
explicitly-requested value.  The difference is that the 'High' version
ensures that the assigned ID is higher than all existing assigned IDs.
For example:

Existing slots  UseNextAvailableSlotID  UseNextAvailableHighSlotID
1-10, 20-30     11                      31

Boot-time assignment of device names can be based on the order of the
slot assignments.  In the above example, the device at slot 20 could get
a *different* name depending on whether there's a device at slot 11 or
not.  This is deemed to be an important issue for network devices
(Client Network Adapters and Trunk Adapters), but not for storage
adapters.  Hence this change set switches to using the 'High' version
for the bld methods of CNA and TrunkAdapter, but not for VStorageAdapter
(the ancestor of V{FC|SCSI}{Client|Server}Adapter).

Change-Id: I65732c4e9a6b17ab1322f562d0ecd449e3b5a588