Added retry exception when parent VIOS is busy
There are certain objects that are children of VIOS.  When the VIOS has
a 'busy' status, queries to these objects will throw a different HTTP
Status.

The children objects are currently the Volume Group and the Network
Bridge.  While the Network Bridge is not directly tied to the VIOS, it
does depend on them being available to work.

This change set adds that new HTTP status code, and appends it to the
default retry logic.

Change-Id: I79762913c0ebf3e8ad7a86eefc991b550f6319e0