Support removal of Arbitrary PVID
The arbitrary PVID is a VLAN ID that gets assigned to additional trunk
adapters within a SEA.  It is a PVID that is used as a placeholder - no
traffic is actually sent upon it.

There are scenarios where an arbitrary PVID may need to be removed from
a network bridge - such that actual traffic could flow upon it perhaps
on a different virtual switch.

This change set adds native support for removing an arbitrary PVID from
the network bridge.

Change-Id: I0b8266c89a8a81b242dddc16d174c85b1357d017