Find lowest trunk/load group for new VLAN
This is the first of a couple change sets to support the load balancing
on network bridges.  The intent of this change set is to find the load
group or trunk adapter that has the lowest count of VLAN ids.

This will ensure that the VLANs get distributed evenly if there are
multiple valid candidates to put the new VLAN on.

Change-Id: I8741dc939d8807b3bb3d3a58122dfa0514e2aabd