WIP: Support NPIV mappings without WWPNs specified
WIP: Unit tests need to be added

The derive_npiv_mappings code required that a predefined set of client
WWPNs be derived.  This is useful for LPM operations, but is not
applicable for initial port mapping build out.

This change set supports the concept of building a port mapping without
requiring the user to predefine the WWPNs.  This is done via a new
derive_base_npiv_mappings method.

Change-Id: Iaf76eea67248d1869d2492fe38f19a722d344076