Support Service Unavai in vios retry helper
This change set adds support in the vios retry helper.  The service
unavailable can occur when a child object of a VIOS can't be returned
because the core VIOS is 'busy'.

The mappings are also not directly clear either.  For example while a
Volume Group is affected by this (and is a direct child of the VIOS), so
is a NetworkBridge (which is not a direct child of a VIOS).  So path
URI inspection is not done here.

While the Service Unavailable could be triggered for other reasons, the
primary culprit is the VIOS busy.  Thus it is folded into here.

Change-Id: I2e8f4f819fdd854fb6f60c8b9ae245ccec6165eb