WIP: Check to support add/remove hw
This change set proposes a method to add or remove hardware (virtual or
physical) to a given system via a single method.

This check could be used for migration as well.

This is being thrown out for a review to see if the general consensus is
that this is a good idea.

Change-Id: I029cc467271e1737b5d2a42ef9136ac2477d5904