VFCMapping lpar_id and find_maps
Previously, the lpar_id getter for a VFCServerAdapter wasn't working
because the field is named differently than for a VSCSIAdapter, but the
getter was being inherited from the generic server adapter base class.
This change set addresses that discrepancy.

It also provides pypowervm.tasks.fc_mapper.find_maps, which corresponds
to pypowervm.tasks.scsi_mapper.find_maps - though it doesn't yet provide
the same level of functionality.

This will be needed as a piece of the effort to clean up storage
artifacts associated with a stale LPAR ID (from an LPAR that has been
deleted).

Change-Id: I5e03d0e86334734922aa6c51aeb590bc73c3e8ec