Task for deleting storage from a VG
Add pypowervm.tasks.storage.rm_vg_storage to remove Virtual Disks and/or
Virtual Optical Media from a Volume Group.

Also adds a stock argmod_func for the retry decorator which refreshes
the Wrapper argument to the decorated function.

This will be part of the procedure to scrub stale storage artifacts from
a deleted LPAR.

Change-Id: I800aae31d9e1de3d4ebf404c2a2f1acdf36a5442