vfc_mapper.find_maps by LPAR UUID
Add the ability to find_maps in a list of VFC mappings by client LPAR
UUID as well as short ID.

This includes refactoring the _id_or_uuid helper to a public method as
pypowervm.utils.uuid.id_or_uuid.

It also adds explicit unit tests for that method and fixes a couple of
bugs uncovered by that effort.

Change-Id: I411ee61366ed3645e63bc1b505b87114bb960b67