Remove WWPN key generation
At one point, the PowerVM API was going to provide a mechanism to
pre-allocate WWPNs on the system.  This never came into existence.  The
pypowervm API had a placeholder that would generate WWPNs (but did so
with random numbers until the API had implemented the function).

Since the API is no longer planning to implement this function,
pypowervm is removing the placeholder function.

Change-Id: I3d572f675f9617c9ce288f80de8dfb618ebf3586