Add retry logic to NPIV parallel mapping
This change set adds a retry decorator to the NPIV mapping action.  It
does a full action retry, because the full operation should be retried.
Even though the updates are running in parallel, this will operate
correctly.  The subsequent reads will gather the previous successful
updates, but then not re-run them.

They will not be re-executed because the mappings would have been seen
as existing already.

Change-Id: Ie767cd325b911cb4c8444a46fe8d1e87bff8526f