Add VIO mapping batcher
Adding and Removing vSCSI/vFC mappings to the VIOS is one of the most
expensive operations for the system.  A typical create/destroy operation
of a LPAR requires several mappings to be added.  If using individual
updates, the requests can take a significant amount of time.

This change set introduces a batching mechanism for VIO updates.  This
is meant to be passed around a series of steps for a LPAR operation.
The mappings that need to be added/removed across many VIOSes will be
documented.

Then upon completion, all of the updates can be run in one single swoop.
For many VIOSes, the requests will be run in parallel (as each VIOS
update can run independently of each other).

The logic will handle the retry and contention of the VIOSes as well.
If a retry is detected, the VIOSes will be refreshed.  Then the mappings
will be added/removed from the object again.

Change-Id: I93b6255181a194d7628954f227f20b6b3b012d97