EntryWrapperGetSpec
We are about to write a decorator that knows how to lock and retry a
transaction on a particular PowerVM object.  We would like the decorated
method (the transaction) to be able to accept *either* an
already-retrieved EntryWrapper *or* its UUID.  Well, its UUID isn't
quite enough; we also need to know its type, and if it's a child, its
parent type/UUID, and any xags.  This is most easily encompassed in a
new class which can know which pieces it needs, and also how to use them
to perform the appropriate GET.  Ultimately, the usage will be:

def some_func(wrapper_or_spec):
  wrapper = wrapper_or_spec.get() if isinstance(
      wrapper_or_spec, EntryWrapperGetSpec) else wrapper_or_spec

The goal is that this should wind up being done only from the
aforementioned lock-load-retry transaction decorator.

Change-Id: I27638bad2f200a28a39f654f669f97c79542aad1