Provide add/remove maps more broadly
As we work towards mechanisms that perform batched updates, we need
granular methods that update the wrappers, but don't actually update the
VIOS.  This is important for an upcoming change set where the wrappers
will be 'batched' into an update pool.

This change set maintains the existing paths for function, but enables
the new mechanism to work as well.

Change-Id: I7376ba64d14b9a1fd65dcc8d1328d10c91f27c8a