@pypowervm.utils.transaction.entry_transaction
A new decorator to facilitate transaction semantics on a PowerVM REST
object.

Example usage:

    @entry_transaction
    def add_gizmos_to_thingy_wrapper(thingy_wrapper, gizmos):
        thingy_wrapper.gizmo_list.extend(gizmos)
        return thingy_wrapper.update()

This decorator provides three things:

1) The decorated method may accept either an EntryWrapper or an
EntryWrapperGetSpec as its first argument.  However, within the body of
the method, that argument is guaranteed to be an EntryWrapper.

2) The decorated method is locked on the UUID of the PowerVM object on
which it operates (represented by its first argument).  Only one method
thus decorated can operate on that PowerVM object at one time.

3) If the decorated method fails due to an etag mismatch - indicating
that the wrapper was updated out-of-band between when it was retrieved
and when it was updated - the wrapper is refreshed and the entire method
is redriven.

Change-Id: I05da76a14301aa4dbed08f3a7760de92f52ac620