Transaction and Subtask
Subtask is an individual unit of work, performed on a wrapper, under
lock.

Transaction bundles Subtasks together and invokes them in sequence under
@entry_transaction (resolve wrapper, lock, retry).

Future: Ability to instantiate a Transaction-esque object with a getter
that defines a Feed (rather than just one Entry) and invokes the
Transaction over each Entry in the Feed in parallel, returning the list
of wrappers, some of which may have been update()d.

Change-Id: Icc8cd7df553f141b220b0bc977fbe9bff4414e30