Add Load Balancing to Network Bridger
This change set adds the concept of Load Balancing of VLANs to the
network bridger.  The logic affects the 'ensure vlan' flow.  It will
detect if load balancing is enabled, but only one 'VEA' (either a Load
Group or Trunk Adapter) is available, we should create a new VEA so that
we start balancing that VLAN across the adapters.

There is 'fuzzy' logic in place to detect if we should just 'fill' up an
existing VEA.  This should be rare and only caused if the administrator
manually modified the VLANs.

There is no logic to rebalance the VLANs across the adapters.

Change-Id: Id5de5b17d2189e900e2cfe9167e53a0ea77a2fe4