WIP: Add transaction manager for transactions
UT TO BE ADDED

Many transactions will need to be run in parallel.  The primary use case
for this will be with the Virtual I/O Server mapping updates.  This
change set adds a very simple transaction Manager that allows for the
creation of several Transactions quickly.

It also handles the complexity of running the Transactions in parallel.
This will replace the existing 'parallel_update' in the util class, but
that will be done in another change set.

Change-Id: I068fa00848c05da7cf4d9e1c72b5e245043cf35a