WIP: FeedTask and FeedGetter
pypowervm.wrappers.entry_wrapper.FeedGetter is a corollary to
EntryWrapperGetter, allowing deferred GET of an entire feed.

This is used to initialize pypowervm.utils.transaction.FeedTask,
a TaskFlow Task which facilitates building up Transactions to be invoked
on each EntryWrapper in a feed.

Also renames pypowervm.utils.transaction.Transaction to EntryTask.

Change-Id: I84ff2ca29bc18f028cc0d83cee31999f3d57b643