validate LPAR proc compat against host proc compat modes
The proc compat mode should be validated across the host's
supported proc compat modes to avoid LPAR deploys with
unsupported proc compat modes from failing in the backend.

If validation fails a ValueError is raised with a message
showing LPAR proc compat is not within the list of host
processor compatability modes.

Change-Id: Idb4c4cdeca3a2bc83aa722ca158a19b84e7cd924