Customizable fixtures (WrapperTaskFx, FeedTaskFx)
In pypowervm.tests.test_fixtures:

Logger: A mixin base class for rudimentary "logging" behavior - simply
appending strings to an ordered list which can be retrieved or reset.

SimplePatcher: A system for simple patching/mocks.

LoggingPatcher: A SimplePatcher whose mock side_effect is to log its
name before returning the configured value.

SimplePatchingFx: A fixture using SimplePatcher.

WrapperTaskFx, FeedTaskFx: Fixtures for
pypowervm.utils.transaction.WrapperTask and FeedTask which make use of
the above to provide default behavior for the primitives of those
classes, but also allow customization on the fly.

Change-Id: I73ed0c9d1bccc676be6fab3e9b555e6257582e07