Add control channel properties to SEA and NetBridge
Expose the ControlChannelInterfaceName as a property on the SEA
object.  Expose the ControlChannelID as a property on the NetBridge
object.

I added code to pvmhttp.py to search the tasks data directory in
addition to the wrappers data directory.  This prevents a need to copy
over the data file, or to use long paths in the testcase itself.

Change-Id: Ic59348bd02808904a9ff2ccf346f33200e49fdc8