WrapperTask/FeedTask allow empty Subtasks
By default, attempting to execute a WrapperTask with no Subtasks will
raise an exception.  This change set adds a boolean allow_empty flag to
the initializer allowing this case to pass instead, returning None.

Previously, attempting to execute a FeedTask containing *any*
WrapperTask with no Subtasks would cause that WrapperTask to fail,
possibly (?) resulting in the failure of the entire flow.  This change
set alters this behavior to pass the aforementioned allow_empty flag to
all WrapperTasks such that empty WrapperTasks are allowed within a
FeedTask.  It also adds a check to ensure that if NO WrapperTasks have
Subtasks, we no-op without even GETting the FeedTask's feed.

Change-Id: I59f699047f1cc476152b1849a80f38bcb34b72fb