Add compute_ltm override to pcm preferences
The compute_ltm (when turned on) overrides the VIOS and network metrics
from coming back to the API.  By default they were turned on.  This is
incorrect behavior and the default should be to keep them turned off.
This change set exposes a field and changes the defaults on the value.

Change-Id: Ie5d99817a14515310eb50fb9d4d9c28d5af4f1ee