Add delay and increase retry count on transactions
The transactions have simplified and increased the concurrency
dramatically when interacting with the system.  But given that there are
no guarantees that the single pypowervm adapter will have the full lock,
the retry logic was included in the transaction.

However, in highly concurrent scenarios it has been found that the retry
logic is perhaps a bit too aggressive.  It retries three times and does
not wait inbetween each retry.

This change set does two things.
- Increases the retries to 5 (up from the 3 default).
- Adds a half second per retry delay.

That second bullet increases the delay by half a second for each retry.
The first retry waits half a second.  The second retry waits a full
second.  Third retry waits 1.5 seconds.  And so on.

This helps reduce a full failure scenario when operations (outside of
the entry_transaction) are operating on an element.

Change-Id: I0f2ecf4e40b2206f421fd2ac7fea426118bf0c1b