Update add_npiv_mappings to return only new maps
This change set modifies the behavior of the vfc_mapper's
add_npiv_port_mappings method to only return the newly added mappings.
It will now ignore the original mappings.

This is important for consumers using multiple fabrics within a given
LPAR.  They need to know which mappings are truly new versus which are
old.

Change-Id: I8a84963a3db2aa9e746474931e3110dc92797a3a
Closes-Bug: 1485099