Order dependence in the client adapter wwpns
The order of the WWPNs is important when interacting with the pypowervm
API.  This was originally not thought to be the case, but this is
important as it dictates the order for how the client WWPNs become
active.

This change set removes the 'set' logic and instead uses 'lists' which
maintain order.

Change-Id: I2921d10af85a8cabea9856f38412f9640a4811ee
Closes-Bug: 1485738