FeedTask, WrapperTask propagate oslo_context
The initializers for FeedTask and WrapperTask now detect the security
context of the invoking thread, an oslo_context.context.RequestContext
instance, and then re-register it in their respective execute methods.
Thus, if a FeedTask and/or WrapperTask is created in one thread, but
executed in another, it will carry its security context with it.

This mechanism facilitates propagation of the parent thread's security
context without intervention by the FeedTask user.

Change-Id: I7360e4e21527da09f66a735cd6f8d1efe9bf4f6b
Closes-Bug: 1487535