WrapperTask accepts oslo_context; FeedTask sets it
The initializer for WrapperTask now accepts an optional 'context'
parameter, which should be an oslo_context.context.RequestContext
instance representing the security context of the parent thread.

FeedTask pulls its context in its main thread (outside of its
WrapperTasks) and propagates it down to the child threads by passing it
into the WrapperTasks.

This mechanism facilitates propagation of the parent thread's security
context without intervention by the FeedTask user.

Change-Id: I7360e4e21527da09f66a735cd6f8d1efe9bf4f6b
Closes-Bug: 1487535