Clean up logging in unit tests
1) Another attempt to resolve the intermittent py34 failure complaining
of inaccurate call count of logging mocks in test_session.py.  This
time, removed the local @mock.patch decorators and instead a) used
test_fixtures.LoggingFx for negative assertions (logger not invoked); b)
used assertLogs for positive assertion (logger invoked).

2) Recent changes in vfc_mapper resulted in an extremely verbose
LOG.warn in a certain code path, which was overwhelming the test output.
Terminated two avians via the single geological expedient of wrapping
that particular test call in assertLogs().  This suppresses the output
in the test run while also confirming that the warning log was invoked.

3) Remove unnecessary logging.basicConfig() from test suites.

Change-Id: Ic9d343146d1af5b7ea1f372c09007ce6938ba2f5