Fix traits in AdapterFx & SessionFx test fixtures
Previously, the AdapterFx and SessionFx test fixtures in
pypowervm.tests.test_fixtures were not defaulting traits - they had to
be specified by the test writer if the test was going to use traits
anywhere in the code path.  The more realistic and user-friendly
mechanism now implemented is to default traits to LocalPVMTraits if not
otherwise specified.

Also fixed the logic that decides which traits to use in various cases
when initializing AdapterFx  They may be specified to AdapterFx(traits)
and/or within AdapterFx(session).  If neither, LocalPVMTraits is used,
as mentioned above.  If both, the traits parameter takes precedence.

Change-Id: I1a87de711a24e8c84694e71daef34eb4f8f82919