Update timeout paradigm
A previous change set increased the timeouts for VIOS updates to 5
minutes as the minimum.  Quickly afterwards, testing against LPARs for
dynamic modifications indicated that some operations in general updates
can take longer than the 60 second default timeout.

The team re-evaluated and determined that create, read and delete
operations should have a timeout of 20 minutes.  While all operations
should be significantly quicker than that, this is a good upper boundary
to indicate that something has failed.

The update has been moved to a broader approach.  All wrapper updates
will now default the timeout to 60 minutes.  There are some significant
DLPAR operations that can take a long amount of time.  This upper
boundary appears sufficient to cover most (if not all) operations and
indicate if something has failed.

Users still have the ability to override timeouts.  But generally this
is not advised as these timeouts are set very high to indicate when
something has gone wrong and the backing API server needs investigation.

Change-Id: Iefb6b5ff49b6c9437c9481c77f6b74b8f747c703