xag arg can be any iterable
Previously, the xag argument to the various REST shim methods needed to
be supplied as a list; some other iterables would cause a failure when
attempting to sort.  This restriction could be painful, for example,
when methods have default arguments, which are not supposed to be
mutable (hence should not be lists).

Now the xag argument can be any iterable.

Change-Id: I768f5b90177e06cb5aa2e9b4e27db6ae53131920