Remove oslo.config from pypowervm
The pypowervm library has sensible defaults across the library.  An
arbitrary config file was never provided to set defaults.

This proposal removes the dependency on oslo.config.  It should also be
noted that oslo.config was never actually in the requirements.txt so it
wouldn't install properly standalone.

Another idea if we agree with this approach is to move all of the
defaults into an __init__.py file.  So that if we find we need to change
the basic defaults, it is at least defined in a single location.

Change-Id: I078acb1d947857b59863c5b2c9f2f340fc8aa1e2