WIP: Add clean_disks option to scrub
TODO: Need to somehow only find orphans if clean_disks = False...

This change set adds a clean_disks option to the storage scrub tasks.
This gets utilized by the discover_hdisk command.

The purpose of this is to provide the user with control as to whether or
not the backing hdisks should be cleaned off the system.  In the case of
discover_hdisk, the invoker may be trying to discover a VM's boot disk
or may simply be trying to discover another volume to attach.

If they're discovering another volume, hit a path where they need to do
a scrub (before the next discover) they could accidentally wipe out the
boot disk.  This change resolves this from occurring.

Change-Id: Ib5825de7a5838ddcf0af682644bf69081d203872