Extend find_maps when no lpar-id is passed
Currently find_maps only returns the matching mapping for the given
client lpar id. Sometimes we need a way to return the matching scsi_mappings
for all the client lpar id for a given backing device.
To support such use cases we have modified the function to have client_lpar_id
as optional.

Change-Id: Ic20c335e9e9de5f3140cc6761eec70f7d07604e3