find_stale_lpars: VSCSI mappings orphaned LPAR IDs
pypowervm.tasks.storage.find_stale_lpars compares the set of client LPAR
IDs across the VSCSI mappings for a given VIOS to the list of LPARs on
that VIOS's host, returning the list of "orphaned" LPAR IDs - that is,
VSCSI mapping client LPAR IDs which don't exist on the host.

Entries in this list can be fed to add_lpar_storage_scrub_tasks to clean
up the mappings (and associated storage) for these LPAR IDs.

Change-Id: Iecd688cfb8247ace113c9ea17688438f46478b06