Use entry_transaction decorator in storage.py
This change is to add the entry_transaction decorator to the
crt_lu() and replace the retry decorator on the rm_ssp_storage()
method.

In cases where we get etag mismatch errors (412) from the REST
layer on the crt_lu(), we were not updating the SSP entry wrapper
and retrying.

Change-Id: I35017c7f050d0dbd1a0e1825d9b7fa9e79d4f376