Support Orphan VLANs
This change set adds support for detection for orphan VLANs on virtual
switches.  An orphan VLAN is one that is used on a Trunk Adapter or
Client Network Adapter on a Virtual I/O Server that is NOT part of a
Network Bridge (more specifically the Shared Ethernet Adapters within a
Network Bridge).

This only impacts the provision flow of new VLANs.  The change will
detect the orphan and raise a very specific error message to help the
user understand how to restore their environment.

The change is written in such a way so that the orphan detection is the
last thing done before a VLAN is provisioned.  It does this to avoid all
the extra calls that go into detected orphan VLANs.

The code does detect this within a given virtual switch.  So if there is
an orphan on ETHERNET0 but the user is provisioning to ETHERNET1, it
will work correctly.

The change also ensures that a new 'arbitrary' VLAN ID does not resolve
to an orphan VLAN id.

Change-Id: I3d08f0c6eb1c0335fd6bc24d58edbf3e3527e350
Closes-Bug: 1493830