Ensure Wrapper subtype to *Getter init
We had a bug deep in the scrub code where a FeedGetter was being
initialized with parent_class=<a string value> instead of a class.  This
was not caught in UT because of mocks.

This change set adds an early check to ensure that the entry_class and
parent_class params to the EntryWrapperGetter and FeedGetter inits are
actually Wrapper subclasses.  With this in place, the offending usage
was caught during UT (and then fixed).

Change-Id: I8055d43506feb9aa8a9f5db147d128f0a8b4caf5
Closes-Bug: 1494845