Include the useful part of a 500 in the exception
Any time a REST call yields a 500 error, the exception message reads:

500 Internal Server Error

...giving no clue as to the actual problem.  The XML payload returned
with the 500 contains a <Message/> element with more details on the root
cause of the problem.  This change set attempts to add that text to the
exception error message of a 500 such that we would see something more
like:

500 Internal Server Error -- The backing device ent4 has a bad
configuration.

Change-Id: I7eb8194238265aa2edc0c2aa34abde714082b38d