Add documentation to the EventHandler
The EventHandler is to be used to manage incoming events from the REST
API.  This may indicate that a URI was created, deleted or modified.
The documentation for how to implement an EventHandler was non-existent.
This adds documentation so that others do not have to learn the API by
diving into the multiple classes used for event processing.

Change-Id: I808a7be9d5b860c4a9516a4cb33e22bbebc12933