Update mapping semantics in migration
This change set tweaks the semantics for the migration mappings to be a
bit more consumable.  Rather than asking the user to build a comma
delimited, slash delimited list, it instead asks for a python list of
inputs for the mappings.  They are still slash delimited, but the format
is easier to understand.

Change-Id: Ieea5785716d6e3a845c4bb22ac72e9b5a4160be8