Add find vios/port by vfc wwpns
This change set adds the ability to find a vios and port by just the
virtual WWPNs.  This is useful in scenarios where the client has
migrated a VM, but perhaps only has the vfc wwpns (and the physical port
that they originally had is no longer valid).  This allows the user to
find the newly hosted physical port on the target server.

Change-Id: Ibedef7ff64a1334ccbfaa27947fe63a95dcee7c4