Simplify and correct vterm code
The vterm code needed simplification.  The run_proc command had improper
documentation and had options that were not used.  It also would throw
errors improperly in the run_proc that were only realized on live
systems.

This change set solves the live systems (something that couldn't be seen
with mocks - as the mocks would make the behavior look correct).  It
resolves the issues with improper documentation and unused parameters
for simplification purposes.

Change-Id: Ibc0e6c0ab04b8dd50c619ad46caac93da048afd6