Don't dump the log stash on 412 (etag mismatch)
pypowervm.helpers.log_helper is designed to save the last several REST
request/response payloads and dump them to the log whenever an exception
is encountered.  In the case of an etag mismatch (HTTP response status
412), this isn't really necessary - we know exactly what happened, and
have most likely coded to retry the operation - and it clutters the
logs.

This change set special-cases the log_helper to avoid dumping the log
stash on an etag mismatch response.

Change-Id: Ia6b50874415e4336033afcc479ac5268596872a8