WIP: Provide 'remotable' local vnc vterm
TODO: Unit Tests
TODO: Live Testing
TODO: validation string, source ip bounding

This change set provides support for a 'remotable' local vnc vterm.  The
standard vnc session is only available to the localhost.  That means
that routing of the VNC must be from a consumer on the localhost.  This
is rare.

This provides a 'repeater'.  This repeater consumes the localhost VNC
and repeats it externally.  This is useful for very specific scenarios -
ex: you want to send the address to perhaps only a single IP.

With this repeater the user can:
 - send the VNC to a specific IP
 - utilize a validation string on connect

This helps ensure that the vterm is only present to the users that
request it.  It is NOT security (as VNC by default is not secure).
Rather it is some additional validation to help ensure that the consumer
is the right user.

Change-Id: I330a04c6ed221a82586ec4c0409ab07a90e69058