EntryWrapper.refresh(use_etag=True), retry w/False
By default, EntryWrapper.refresh uses the etag of the wrapper to avoid
refetching and rewrapping the object if it hasn't changed on the server.
However, there are situations where a consumer may have made local
changes to the wrapper, but then need to restore it to its previous
incarnation.  Doing a default .refresh() in this situation will not
achieve this result because the etag hasn't changed.

This change set introduces an optional parameter to EntryWrapper.refresh
called use_etag.  The default value, True, yields the legacy behavior.
But setting this parameter to False will perform the GET *without* the
etag, thus ensuring that the object is retrieved afresh from the server.

One code path where refresh was being used was in the refresh_wrapper
argmod_func used for @retry.  In the default usage, which comprises
most of the current uses of @retry, the retry is prompted by an etag
mismatch, so it makes no difference whether we send the existing etag or
not, since it shouldn't match.  However, it is possible for @retry to be
triggered by other mechanisms, in which case the default refresh - which
would essentially be a no-op - could leave the wrapper in a state where
it doesn't actually match the server (because local changes have been
made).  To avoid this problem, refresh_wrapper is now calling
EntryWrapper.refresh with the new use_etag argument set to False.

Change-Id: I7f57dd0a5c4d7cc8ed1274ea9f05ebd2bed07e18