_image_lu_in_use tolerates orphan LU linked clones
It is possible for an SSP to get into a bad state where an image LU is
removed even though it is supposed to be backing one or more linked
clones.  In this case, rm_ssp_storage can fail - even for unrelated
LUs - with a stack trace like:

  File "/opt/stack/pypowervm/pypowervm/tasks/storage.py", line 597, in
        rm_ssp_storage
    if _rm_lus(ssp_wrap, lus, del_unused_images=del_unused_images):
  File "/opt/stack/pypowervm/pypowervm/tasks/storage.py", line 569, in
        _rm_lus
    if _image_lu_in_use(ssp_wrap, backing_image):
  File "/opt/stack/pypowervm/pypowervm/tasks/storage.py", line 383, in
        _image_lu_in_use
    if lu.cloned_from_udid[2:] == image_udid:
TypeError: 'NoneType' object has no attribute '__getitem__'

This change set resolves the issue, tolerating image-less linked clones
and logging a warning.

Change-Id: Ief8ea12034620b80e593c994d23d3169683c4eae